+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to
 
 - Enable project level concurrency limit
   [#2906](https://github.com/OpenFn/lightning/issues/2906)
+- Transfer credentials ownership to a project collaborator
+  [#2820](https://github.com/OpenFn/lightning/issues/2820)
 - Delete unused snapshots on workorders retention cleanup
   [#1832](https://github.com/OpenFn/lightning/issues/1832)